fixes for feeds/recent Minor fix for tag interface perltidied controllers.

+++ lib/MojoMojo/C/Jsrpc.pm
@@ -25,18 +25,17 @@ params->{content} and runs it through the formatter chain.
 =cut
 
 sub render : Path('/.jsrpc/render') {
-        my ( $self, $c ) = @_;
-        my $output="Please enter something";
-        if ($c->req->params->{content} &&
-            $c->req->params->{content} =~/(\S+)/)
-        {
-            $output= MojoMojo::M::Core::Content->formatted
-            (
-             $c->req->base,
-             $c->req->params->{content},
-            );
-        }
-        $c->res->output($output);
+    my ( $self, $c ) = @_;
+    my $output = "Please enter something";
+    if (   $c->req->params->{content}
+        && $c->req->params->{content} =~ /(\S+)/ )
+    {
+        $output =
+          MojoMojo::M::Core::Content->formatted( $c->req->base,
+            $c->req->params->{content},
+          );
+    }
+    $c->res->output($output);
 }
 
 =item diff (/.jsrpc/diff) 
@@ -47,25 +46,26 @@ and diffs it against the previous version.
 =cut
 
 sub diff : Path('/.jsrpc/diff') {
-        my ( $self, $c, $revision ) = @_;
-        $revision= MojoMojo::M::Core::Revision->retrieve($revision);
-        if (my $previous=$revision->previous) {
-        $c->res->output($revision->formatted_diff($c->req->base,
-                                                  $previous));
-        } else {
+    my ( $self, $c, $revision ) = @_;
+    $revision = MojoMojo::M::Core::Revision->retrieve($revision);
+    if ( my $previous = $revision->previous ) {
+        $c->res->output(
+            $revision->formatted_diff( $c->req->base, $previous ) );
+    }
+    else {
         $c->res->output("This is the first revision!");
-      }
+    }
 }
 
 =item submittag (/.jsrpc/submittag)
 
-Add a path through submit
+Add a tag through form submit
 
 =cut
 
 sub submittag : Path('/.jsrpc/submittag') {
     my ( $self, $c, $page ) = @_;
-    $c->req->args([$page,$c->req->params->{tag}]);
+    $c->req->args( [ $page, $c->req->params->{tag} ] );
     $c->forward('/jsrpc/tag');
 }
 
@@ -76,18 +76,27 @@ add a tag to a page. return list of yours and popular tags.
 =cut
 
 sub tag : Path('/.jsrpc/tag') {
-        my ( $self, $c, $page, $tag ) = @_;
-        $page = MojoMojo::M::Core::Page->retrieve($page);
-        $c->log->_dump($page);
-        unless (MojoMojo::M::Core::Tag->search(page=>$page,
-                                            person=>$c->req->{user_id},
-                                            tag=>$tag)->next()) {
-          $page->add_to_tags({tag=>$tag,
-                              person=>$c->req->{user_id}}) 
-                              if $page;
-        }
-        $c->req->args([$page]);
-        $c->forward('/page/tags');
+    my ( $self, $c, $page, $tagname ) = @_;
+    $page = MojoMojo::M::Core::Page->retrieve($page);
+    $c->log->_dump($page);
+    unless (
+        MojoMojo::M::Core::Tag->search(
+            page   => $page,
+            person => $c->req->{user_id},
+            tag    => $tagname
+        )->next()
+      )
+    {
+        $page->add_to_tags(
+            {
+                tag    => $tagname,
+                person => $c->req->{user_id}
+            }
+          )
+          if $page;
+    }
+    $c->req->args( [ $page, $tagname ] );
+    $c->forward('/page/tags');
 
 }
 
@@ -98,15 +107,18 @@ remove a tag to a page. return list of yours and popular tags.
 =cut
 
 sub untag : Path('/.jsrpc/untag') {
-        my ( $self, $c, $page,$tag ) = @_;
-        ($page) = MojoMojo::M::Core::Page->retrieve($page);
-        $tag=MojoMojo::M::Core::Tag->search(page=>$page,
-                                            person=>$c->req->{user_id},
-                                            tag=>$tag)->next();
-        $tag->delete() if $tag;
-        $c->req->args([$page]);
-        $c->forward('/page/tags');
-    }
+    my ( $self, $c, $page, $tagname ) = @_;
+    $page = MojoMojo::M::Core::Page->retrieve($page);
+    die "Page " . $page . " not found" unless ref $page;
+    my $tag = MojoMojo::M::Core::Tag->search(
+        page   => $page,
+        person => $c->req->{user_id},
+        tag    => $tagname
+    )->next();
+    $tag->delete() if $tag;
+    $c->req->args( [ $page, $tagname ] );
+    $c->forward('/page/tags');
+}
 
 =head1 AUTHOR
